Fix discount error counting.

+++ server/shoppingCart.js
@@ -141,7 +141,7 @@ module.exports = class {
             var command = "SELECT event.Discount_Rate, event.Event_Category, product.Price\
                             FROM event, order_main, order_content, product\
                             WHERE order_main.Customer = '" + req.body.customer + "' AND order_content.Order_Number = order_main.OID AND \
-                                product.PID = order_content.Item AND product.Category = event.Event_Category";
+                                product.PID = order_content.Item AND product.Category = event.Event_Category AND order_main.Status = 0";
 
             var db = DataBaseController.GetDB();
             var discount = 0;
@@ -167,9 +167,14 @@ module.exports = class {
                 res.send("Deletion success!");
             }
             var callback = function(error, rows, fields) {
-                var delCommand = "DELETE FROM order_content\
-                            WHERE order_content.Order_Number = " + rows[0].OID;
-                db.query(delCommand, clear);
+                if (rows.length == 0) {
+                    res.send("");
+                }
+                else {
+                    var delCommand = "DELETE FROM order_content\
+                                WHERE order_content.Order_Number = " + rows[0].OID;
+                    db.query(delCommand, clear);
+                }
             }
 
             db.query(command, callback);
